Integrate all the variants
